dodanie zamowienia

+++ src/main/java/com/packt/webstore/domain/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package com.packt.webstore.domain.repository;
 import com.packt.webstore.domain.Order;
 
 public interface OrderRepository {
-	Long saveOrder(Order order);
+	Integer saveOrder(Order order);
 }
